Moved thumbnail controllers to ViewerThumbnails file to facilitate more consistent positioning
