+++ client/src/components/ProductSubs/ProductImages.jsx
@@ -32,26 +32,6 @@ const LeftArrow = styled.div`
   &:hover {color: #80ccc4; };
 `;
 
-const UpArrow = styled.div`
-  color: #e0e0e0;
-  font-size: 1rem;
-  z-index: 12;
-  position: absolute;
-  top: 2.75vh;
-  left: 16vw;
-  &:hover {color: #80ccc4; };
-`;
-
-const DownArrow = styled.div`
-  color: #e0e0e0;
-  font-size: 1rem;
-  z-index: 12;
-  position: absolute;
-  top: 50.25vh;
-  left: 16vw;
-  &:hover {color: #80ccc4; };
-`;
-
 const Image = styled.img`
   width: 100%;
   height: 100%;
@@ -66,11 +46,8 @@ class ProductImages extends React.Component {
     super(props);
     this.state = {
       index: 0,
-      start: 0,
-      end: 4,
     };
     this.clickNavHandler = this.clickNavHandler.bind(this);
-    this.clickThumbNavHandler = this.clickThumbNavHandler.bind(this);
     this.clickedThumb = this.clickedThumb.bind(this);
   }
 
@@ -82,17 +59,6 @@ class ProductImages extends React.Component {
     }
   }
 
-  clickThumbNavHandler(event) {
-    const direction = Number(event.target.id);
-    if (this.state.start + direction >= 0
-      && this.state.end + direction < this.props.images.length) {
-      this.setState((prevState) => ({
-        start: prevState.start += direction,
-        end: prevState.end += direction,
-      }));
-    }
-  }
-
   clickedThumb(index) {
     this.setState({ index: Number(index) });
   }
@@ -101,8 +67,6 @@ class ProductImages extends React.Component {
     return (
       <Wrapper>
         <ViewerThumbnails start={this.state.start} end={this.state.end} images={this.props.images} clickedThumb={this.clickedThumb} id={this.props.id} alt="" />
-        <UpArrow onClick={this.clickThumbNavHandler}><FontAwesome id="-1" name="angle-up" size="2x" /></UpArrow>
-        <DownArrow onClick={this.clickThumbNavHandler}><FontAwesome id="1" name="angle-down" size="2x" /></DownArrow>
         <RightArrow onClick={this.clickNavHandler}><FontAwesome id="1" name="angle-right" size="2x" /></RightArrow>
         <LeftArrow onClick={this.clickNavHandler}><FontAwesome id="-1" name="angle-left" size="2x" /></LeftArrow>
         <Image src={this.props.images[this.state.index].url} key={this.props.id} alt="style photograph" />
